Polymarket Trader/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==13.1
pydantic==2.10.4
httpx==0.27.2
python-multipart==0.0.20

# Async utilities
asyncio-throttle==1.0.2
aiohttp==3.11.11

# LLM / AI
openai==1.59.9
ollama==0.4.4

# Web search (optional — only needed if you have a Tavily API key)
# tavily-python==0.5.0

# Polymarket SDK & crypto
py-clob-client==0.17.0
eth-account==0.13.4

# Data & utilities
python-dotenv==1.0.1
configparser==7.1.0
python-dateutil==2.9.0
pydantic-settings==2.7.0

# Concurrency
anyio==4.7.0
